Improvement: Guests use both sides of wide paths (Work in Progress)
Working on an idea for a way to spread peeps out on wide paths. Instead of only one wide flag, each path now has 8. When the Map updates the path wide flags it does so in 8 different orders, with each order being assigned their own wide flag.

When peeps are path finding, the flag they consider when checking if
a path is wide is based on their ID. The isWide flag is still keeps track of the default update order. Staff peeps use the default is wideFlag.

todo:
- replace arrays with structs

- replace edge/connection/direction variables with direct calls

- add new flags to save file (?)

- add way for debugger to view all wideFlags