Improvement: Guests use both sides of wide paths (Work in Progress)
Working on an idea for a way to spread peeps out on wide paths. Instead of only one wide flag, each path now has eight. When the map updates the path wide flags it does so in two directions for each corner of the map. Each of these update orders is assigned their own wide flag for each footpath.

The function to update the wide flag now uses edge/connection/direction relative to the new starting tile and update direction.

When peeps are path finding, the flag they consider when checking if a path is wide is based on their ID. The original wide flag still keeps track of the default update order. Staff peeps use the default wide Flag.

A stop gap is currently in place to view the flags with the debugger by cycling through all eight groups each month.

todo:

-replace arrays with structs

-replace edge/connection/direction variables with direct calls

-add new flags to save file (?)

-add better way for debugger to view all wideFlags

-find better names for constants/variables