Improvement: Guests use more of wide paths (Work in Progress)
Working on an idea for a way to spread peeps out on wide paths. Instead of only one wide flag, each path now has eight. When the map updates the path wide flags it does so in two directions for each corner of the map. Each of these update orders is assigned their own wide flag for each footpath.

The function to update the wide flag now uses edge/connection/direction relative to the new starting tile and update direction.

When peeps are path finding, the flag they consider when checking if a path is wide is based on their ID. The original wide flag still keeps track of the default update order. Staff peeps use the default wide Flag.

Each group of flags can be viewed with the debugger.

ToDo:

-find way for flags not to be reliant on taking up space in current structures. (?)

-add new flags to save file (?)

Also considering:

-Adding 8 more groups that use the centre of paths that are wide to everyone so far