Fix: x button in the object selector window
-Pressing the x for the object selection window in scenario editor will now
close the object selection window and advance it to landscape editor
instead of closing the entire scenario editor and sending the user back to
the main menu.

-Clicking the top right x for the object selection window in track designer
will now close the object selection window and advance it to track design
instead of closing the entire track designer and sending the user back to
the main menu.